+++ examples/redux/lib/reducer/AppReducer.p.dart
@@ -11,5 +11,5 @@ class AppReducerProvider
     extends Provider<AppState, AppStateAction, AppReducer> {
   AppReducerProvider.createWith(AppReducer reducer) : super(reducer);
   AppReducerProvider() : super(AppReducer());
-  static const key = '2c79ffad-1ebd-4c88-9007-56f6c391d67f';
+  static const key = '7bae86e0-fc83-447e-afba-aea7a3e9e879';
 }