+++ packages/redux/generator/lib/builders/config/FileExtensions.dart
@@ -1,6 +1,6 @@
 class FileExtensions {
   FileExtensions._internal();
 
-  static const Provider = '.provider.dart';
-  static const Store = '.redux.dart';
+  static const Provider = '.g.part';
+  static const Store = '.p.dart';
 }