+++ packages/redux/CHANGELOG.md
@@ -4,6 +4,11 @@ Implement Redux Generator:
   * Middleware
   * Save state
 
+## 0.0.2
+Implement Redux Generator:
+  * Remove Lazy (All is Lazy)
+  * *.provider.dart & *.store.dart => *.p.dart
+
 ## 0.0.1
 Implement Redux Generator:
   * Reducer Provider